app: 单段位置模式运动 + 速度模式 + 弧线路徑
- Mecanum_Move: 新增 Mecanum_MoveWithEncoder 单段位置模式（替代两段式斜坡）
  速度模式 API: Mecanum_VelocityMove / Mecanum_VelocityStop / Mecanum_EnableAll
  弧线路徑 Mecanum_RunPath（5段: 前进/左转/前进/侧移/半圆弧）
  MECANUM_OMEGA_SIGN 修正旋转方向
- NavigationMecanum: Nav_GoToWorld 改用单段位置模式，移除 g_nav_ramp
- 到位检测: osDelay(duration_s) + 1s轮询 Emm_V5_Is_Reached

+++ STM32G4_banyunche/app/NavigationMecanum.h
@@ -40,7 +40,7 @@ extern World_Dir Self_Dir;
  * yaw 可使用 MECANUM_DEG_TO_RAD 辅助书写，例如 90.0f * MECANUM_DEG_TO_RAD
  */
 extern World_Dir g_waypoints[NAV_WAYPOINT_MAX];
-extern int      g_waypoint_count;
+extern uint8_t      g_waypoint_count;
 
 /* ============================================================
  * 函数声明
